% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verb-fit.R
\name{.fit_status_vector}
\alias{.fit_status_vector}
\title{0/1 event vector of a fit's training data}
\usage{
.fit_status_vector(fit)
}
\description{
0/1 event vector of a fit's training data
}
\keyword{internal}
