% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verb-interpret.R
\name{.interpret_dispatch}
\alias{.interpret_dispatch}
\title{Adaptors from interpret() method ids to compute_*() signatures}
\usage{
.interpret_dispatch()
}
\description{
Adaptors from interpret() method ids to compute_*() signatures
}
\keyword{internal}
