% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling.R
\name{.make_splits}
\alias{.make_splits}
\title{Expand a resampling specification into concrete row-index splits}
\usage{
.make_splits(resampling, data, status_col = NULL)
}
\arguments{
\item{resampling}{A \code{survalis_resampling} object.}

\item{data}{The (already row-filtered) modelling data frame.}

\item{status_col}{Name of the event-indicator column, used when
\code{params$strata == "status"}.}
}
\value{
A list of \verb{list(analysis = <int>, assessment = <int>, id = <chr>)}.
}
\description{
Expand a resampling specification into concrete row-index splits
}
\keyword{internal}
