% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_functions.R
\name{list_tunable_survlearners}
\alias{list_tunable_survlearners}
\title{List tunable survival learners}
\usage{
list_tunable_survlearners()
}
\value{
A base \code{data.frame} like \code{list_survlearners()} but containing only
rows where \code{tune} is not \code{NA}.
}
\description{
Filters learners to only those that provide a \verb{tune_*} function in addition
to \verb{fit_*} and \verb{predict_*}.
}
\examples{
list_tunable_survlearners()
}
