% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survalis-package.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{Surv}
\title{Re-export Surv from survival}
\examples{
Surv(c(5, 8, 12), c(1, 0, 1))
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{survival}{\code{\link[survival]{Surv}}}
}}

