% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_rpart.R
\name{fit_rpart}
\alias{fit_rpart}
\title{Fit a Survival Tree Model using \code{rpart}}
\usage{
fit_rpart(
  formula,
  data,
  minsplit = 20,
  minbucket = round(minsplit/3),
  cp = 0.01,
  maxcompete = 4,
  maxsurrogate = 5,
  usesurrogate = 2,
  xval = 10,
  surrogatestyle = 0,
  maxdepth = 30
)
}
\arguments{
\item{formula}{A survival formula of the form \code{Surv(time, status) ~ predictors}.}

\item{data}{A \code{data.frame} containing the variables in the model.}

\item{minsplit}{Minimum number of observations that must exist in a node in order for a split to be attempted.
Default is \code{20}.}

\item{minbucket}{Minimum number of observations in any terminal node. Default is \code{round(minsplit / 3)}.}

\item{cp}{Complexity parameter for pruning. Default is \code{0.01}.}

\item{maxcompete}{Number of competitor splits retained in the output. Default is \code{4}.}

\item{maxsurrogate}{Number of surrogate splits retained in the output. Default is \code{5}.}

\item{usesurrogate}{How surrogates are used in the splitting process. Default is \code{2}.}

\item{xval}{Number of cross-validations to perform in \code{rpart}. Default is \code{10}.}

\item{surrogatestyle}{Controls selection of surrogate splits. Default is \code{0}.}

\item{maxdepth}{Maximum depth of any node of the final tree. Default is \code{30}.}
}
\value{
An object of class \code{"mlsurv_model"} containing:
\itemize{
\item \code{model} -- fitted \code{rpart} survival tree object
\item \code{learner} -- \code{"rpart"}
\item \code{formula}, \code{data}, \code{time}, and \code{status}
}
}
\description{
Fits a survival tree using the \pkg{rpart} package with an exponential splitting rule.
This learner is compatible with the \code{mlsurv_model} interface and can be used with
the \code{cv_survlearner()} and \verb{tune_*()} functions in the \pkg{survalis} framework.
}
\details{
This function fixes the \code{method = "exp"} argument to fit an exponential
splitting survival tree, which models the hazard function assuming
exponential survival within terminal nodes.
}
\examples{
mod_rpart <- fit_rpart(Surv(time, status) ~ age + karno + celltype, data = veteran)
pred_rpart <- predict_rpart(mod_rpart, newdata = veteran[1:5, ], times = c(100, 200, 300))
head(pred_rpart)

}
\references{
Therneau TM, Atkinson EJ. (2019). \emph{An Introduction to Recursive Partitioning Using the RPART Routines}. Mayo Clinic.
}
