% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ranger.R
\name{predict_ranger}
\alias{predict_ranger}
\title{Predict Survival Probabilities from a ranger Model}
\usage{
predict_ranger(object, newdata, times)
}
\arguments{
\item{object}{An \code{"mlsurv_model"} object returned by \code{\link{fit_ranger}}.}

\item{newdata}{A \code{data.frame} containing the same predictors as the training data.}

\item{times}{Numeric vector of time points at which to estimate survival probabilities.}
}
\value{
A \code{data.frame} with one row per observation in \code{newdata}
and one column per time point (columns named \code{"t=<time>"}).
}
\description{
Generates predicted survival probabilities for given time points
from a model fitted with \code{\link{fit_ranger}}.
}
\details{
Predictions are obtained from \code{\link[ranger]{predict.ranger}} and
survival curves are interpolated to match the requested \code{times}.
}
\examples{
mod <- fit_ranger(Surv(time, status) ~ age + karno + celltype, data = veteran)
predict_ranger(mod, newdata = veteran[1:5, ], times = c(100, 200, 300))

}
\seealso{
\code{\link{fit_ranger}}, \code{\link{tune_ranger}}
}
