% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survalis-package.R
\docType{package}
\name{survalis-package}
\alias{survalis}
\alias{survalis-package}
\title{SurvALIS: Interpretable Survival Machine Learning}
\description{
Core learners, tuning, evaluation, and interpretability utilities for survival analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://CRAN.R-project.org/package=survalis}
  \item Report bugs at \url{https://github.com/ielbadisy/survalis/issues}
}

}
\author{
\strong{Maintainer}: Imad El Badisy \email{elbadisyimad@gmail.com}

}
\keyword{internal}
