% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survalis-package.R
\docType{package}
\name{survalis-package}
\alias{survalis}
\alias{survalis-package}
\title{SurvALIS: Interpretable Survival Machine Learning}
\description{
Core learners, tuning, evaluation, and interpretability utilities for survival analysis.
}
\author{
\strong{Maintainer}: Imad EL BADISY \email{elbadisyimad@gmail.com} [copyright holder]

}
\keyword{internal}
